requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
python-dotenv
pydantic-settings
